feat/cashier: add cashierRequest & cashierResponse in web package
