+++ models/web/response.go
@@ -11,15 +11,3 @@ type AdminResponse struct {
 	FullName string `json:"fullname"`
 	Username string `json:"username"`
 }
-
-type CashierLoginResponse struct {
-	Username string `json:"username"`
-	Token string `json:"token"`
-}
-
-type CashierResponse struct {
-	ID uint `json:"id"`
-	AdminID uint `json:"adminID"`
-	Fullname string `json:"fullname"`
-	Username string `json:"username"`
-}
